==> Greek/05. Inference.ipynb <==
import pickle
import spacy

# Load Label encoder
encoder = pickle.load(open('Model/Label_encoder.pkl', 'rb'))
# Load prediction model
model = pickle.load(open('Model/model.pkl', 'rb'))
# Load spacy model
nlp = spacy.load("el_core_news_sm")

def inference(text):
    # Get embeddings
    embeddings = nlp(text).vector
    # Get model's prediction
    pred = model.predict( embeddings.reshape(1,-1) )
    # Return class
    return encoder.inverse_transform(pred)


inference('Νίκη')
